# src/main/resources/openapi.yaml
openapi: "3.0.3"
info:
  title: Users Service API
  description: REST API for profiles, settings, and avatars for the chat application.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local Users Service
  - url: http://localhost:8080/api
    description: Through API Gateway (prefix /api)
tags:
  - name: Users
    description: User profile and settings endpoints
  - name: Media
    description: Avatar and media-related endpoints
paths:
  /users/health:
    get:
      tags: [Users]
      summary: Health check
      operationId: health
      security: []
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: OK

  /users/me/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: User identifier; either a userId (UUID/string from JWT sub) or a username
        schema:
          type: string
    get:
      tags: [Users]
      summary: Get current user's private profile
      operationId: getMe
      responses:
        '200':
          description: The current user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: User identifier; either a userId (UUID/string from JWT sub) or a username
        schema:
          type: string
    get:
      tags: [Users]
      summary: Get a public profile by userId or username
      operationId: getPublicUser
      responses:
        '200':
          description: Public profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/me/avatar:
    post:
      tags: [Media]
      summary: Request a pre-signed URL to upload an avatar
      operationId: requestAvatarUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignRequest'
            examples:
              pngExample:
                value:
                  contentType: image/png
                  filename: avatar.png
      responses:
        '200':
          description: Pre-signed upload information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /users/create:
    post:
      tags: [ Users ]
      summary: Create new user if doesn't exist
      operationId: createNewUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'

      responses:
        '200':
          description: Pre-signed upload information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /users/update/{userId}:
      parameters:
        - name: userId
          in: path
          required: true
          description: User identifier; either a userId (UUID/string from JWT sub) or a username
          schema:
            type: string
      post:
        tags: [ Users ]
        summary: Updates user profile
        operationId: updateUser
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        responses:
          '200':
            description: Updated user profile
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/PresignResponse'
          '400':
            $ref: '#/components/responses/BadRequest'
          '401':
            $ref: '#/components/responses/UnauthorizedError'

  /users/send/notification/{senderuuId}:
    parameters:
      - name: senderuuId
        in: path
        required: true
        description: User identifier; either a userId (UUID/string from JWT sub) or a username
        schema:
          type: string
    post:
      tags: [ Notification ]
      summary: Send contact notification
      operationId: sendContactNotification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'

      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  responses:
    UnauthorizedError:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                status: 401
                error: Unauthorized
                message: Bearer token required
    BadRequest:
      description: The request payload is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    User:
      type: object
      required: [userId, username, displayName]
      properties:
        userId:
          type: string
          description: Stable identifier (typically from JWT `sub`).
          example: abc-123
        username:
          type: string
          example: user_abc123
        displayName:
          type: string
          example: Alex
        bio:
          type: string
          nullable: true
          maxLength: 200
        avatarUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        contacts:
          type: array
          items:
            allOf:
            - $ref: '#/components/schemas/Contacts'

    Contacts:
      type: object
      required: [ userId, username, displayName ]
      properties:
        userId:
          type: string
        username:
          type: string
        bio:
          type: string
        displayName:
          type: string

    Notification:
      type: object
      required: [ senderUsername, userId, receiverUsername ]
      properties:
        userId:
          type: string
        senderUsername:
          type: string
        receiverUsername:
          type: string

    UserPublic:
      allOf:
        - $ref: '#/components/schemas/User'
      description: Public view of a user profile; sensitive/private fields omitted by the service.
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 40
        bio:
          type: string
          maxLength: 200
        avatarUrl:
          type: string
          format: uri
      example:
        displayName: "Alex"
        bio: "Grappler | Java dev"

    PresignRequest:
      type: object
      required: [contentType]
      properties:
        contentType:
          type: string
          description: MIME type of the avatar to upload
          example: image/png
        filename:
          type: string
          description: Optional client-side filename hint
          example: avatar.png
    PresignResponse:
      type: object
      required: [uploadUrl, publicUrl, method, headers]
      properties:
        uploadUrl:
          type: string
          format: uri
          description: URL to PUT the avatar content to
        method:
          type: string
          example: PUT
        headers:
          type: object
          additionalProperties:
            type: string
          description: Headers required by the storage provider (e.g. Content-Type)
        publicUrl:
          type: string
          format: uri
          description: Public URL the service will store as avatarUrl after upload
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string